requirements.txt
streamlit
google-generativeai
crewai
crewai-tools
langchain-google-genai
pysqlite3-binary